feat: implement epics 3 and 4 - book management, assignments, and activity players
This comprehensive commit implements multiple epics and stories for the Dream LMS platform.

Epic 3: Book Management & Teacher Tools
========================================

Story 3.0: Dream Central Storage Setup & Configuration
- Implement DCS API client with authentication
- Add webhook registration service for book sync
- Add configuration parser for DCS settings

Story 3.1: Dream Central Storage API Integration
- Implement book sync from DCS
- Add webhook endpoints for book updates
- Add book service layer

Story 3.2: Book Catalog & Activity Data Models
- Add Book and Activity models
- Add database migrations for book tables
- Add BookAccess for publisher permissions

Story 3.3: Book Asset Proxy & Delivery
- Implement asset proxy endpoint
- Add image URL transformation for DCS assets

Story 3.5: Teacher Class Management
- Add Class model and API endpoints
- Implement class-student relationships
- Add class creation and management

Story 3.6: Book Catalog Browsing for Teachers
- Implement books API with filtering
- Add activity listing by book

Story 3.7: Assignment Creation Dialog & Configuration
- Add multi-step assignment creation dialog
- Implement assignment configuration (due date, time limit, recipients)
- Add assignment API endpoints

Story 3.8: Teacher Assignment Management Dashboard
- Add assignment listing for teachers
- Add edit and delete assignment dialogs
- Implement assignment statistics

Story 3.9: Student Assignment View & Dashboard
- Add student assignment listing
- Filter by status (not_started, in_progress, completed)
- Add assignment detail view with "Start" button

Epic 4: Activity Players & Submission
======================================

Story 4.1: Activity Player Framework & Layout
- Create ActivityPlayerLayout with header/footer
- Add ActivityTimer component
- Implement player infrastructure for all activity types

Story 4.2: Fix Activity Players with Real Config
- Fix CirclePlayer, DragDropPicturePlayer, MatchTheWordsPlayer
- Add DragDropPictureGroupPlayer
- Fix PuzzleFindWordsPlayer
- Parse real activity config from DCS

Story 4.7: Assignment Submission & Result Storage
- Implement submission API endpoint
- Add useAssignmentSubmission hook
- Create success page with score display
- Add scoring calculation logic

Backend Changes:
================
Models:
- Add Book, Activity, BookAccess, Class, ClassStudent models
- Update Assignment and AssignmentStudent models
- Add webhook event logging

API Routes:
- /api/books - Book listing and filtering
- /api/book-assets - Asset proxy
- /api/classes - Class management
- /api/students - Student operations
- /api/assignments - Assignment CRUD and start
- /api/webhooks - DCS webhook handling

Services:
- DreamStorageClient - DCS API integration
- BookService - Book sync and management
- ConfigParser - DCS configuration parsing
- WebhookRegistration - Auto-register webhooks

Tests:
- 35+ unit tests for API endpoints
- Integration tests for book sync and DCS
- Test fixtures for all models

Frontend Changes:
=================
Components:
- Assignment creation wizard (5-step dialog)
- Book browsing and activity listing
- Activity players for all 6 activity types
- Assignment management (edit, delete)
- Success page with confetti animation

Hooks:
- useAssignmentSubmission - Handle submission flow
- useActivityTimer - Track time spent
- useWindowSize - Responsive layout

Services:
- assignmentsApi - Assignment API client
- booksApi - Books API client
- teachersApi - Teacher API client

Types:
- Comprehensive TypeScript types for Assignment, Book, Activity

Infrastructure:
- Environment configuration
- Asset handling
- Navigation context
- Route generation

QA & Documentation:
===================
- Quality gates for stories 3.0-3.9, 4.7
- Comprehensive story documentation
- Dev agent records for all stories
- Test coverage reports

+++ backend/app/models.py
@@ -773,7 +773,7 @@ class AssignmentStatus(str, Enum):
 class AssignmentStudentBase(SQLModel):
     """Shared AssignmentStudent properties"""
     status: AssignmentStatus = Field(default=AssignmentStatus.not_started)
-    score: int | None = Field(default=None, ge=0, le=100)
+    score: float | None = Field(default=None, ge=0, le=100)
     answers_json: dict | None = Field(default=None, sa_column=Column(JSON))
     progress_json: dict | None = Field(default=None, sa_column=Column(JSON))
     started_at: datetime | None = Field(default=None)
